Allow for more compact display of image collection tilesets
It may still not be ideal, but instead of using the maximum tile size of
the entire collection for each tile it will now use the maximum width
and height of only those tiles in the same column and row respectively.

This solution is only used when compiling against Qt 5.2 or higher,
since older Qt versions only take into account visible cells so the
display breaks when larger tiles become visible by scrolling.