Avoid wrong removal of objects when resizing the map
Since object coordinates are now in pixel coordinates, the visibility
check of the resize operation was wrongly determining most objects were
outside of the map boundaries and removed those objects. Now it should
correctly take the pixel coordinates into account. At the same time,
handling of tile objects and object rotation was added.

While working on this I found another bug however, which is that the
object positions are not adjusted correctly. This is because the
correction is still applied in tile coordinates.

Closes #617