Removed a non-functional hack that was causing a crash
This code was meant to make sure positions and sizes of objects stored
as part of tile collision information were in coordinates relative to
the tile texture, ranging from 0-1. However, it is no longer functional
since the map's tile width and height are no longer used by the map
reader and writer now that the objects store their positions in pixels
rather than tiles.

Overall I think it's anyway a better idea when the positions and sizes
are kept in pixels here and not stored in "texture coordinates", since
this is consistent with the units used in regular object groups. In
addition it is trivial to divide the values by the size of the tile to
obtain the texture coordinates.

Fixes #646