Fixed another possible crash in the orthogonal renderer
This basically reverts the incomplete 5f6b5cadbd436183a5c530b868d and
fixes the problem a little more reliably.

Starting point only needs to be bound to the first tile of the layer and
ending point only needs to be bound to the last tile. As long as we do
nothing when the starting point ends up beyond the ending point.