Fixed Windows installer again
This reverts back to the old behavior, which is also broken but it's
less broken than without this change.

The problem is that _?= is setting the installation directory and
thereby breaking the uninstall if it is supposed to happen from a
different directory than where Tiled is being installed. However,
omitting it breaks ExecWait, which causes the uninstall to run parallel
to the install process, deleting files that have just been installed and
leaving a broken installation.

So until there is a proper fix, I'm reverting back to the previous
behavior.