qmake: Added project files to run Tiled Quick on Android
Seems to mostly work, though it did seem to crash for me after a bit of
scrolling and the file picker that pops up is really ugly.

To keep deployment simple, the source code of the Tiled Quick QML plugin
and libtiled are simply compiled into the executable. The Tiled Quick UI
is included using a Qt resource file.