Added an auto layer function.
When you enable auto layer it will look at the tile set name and try to direct the tiles to the appropriate layer based on the extension.
E.G. If you have 2 layers one named t1 and the other named t2 and two tile sets named grassland.t1 and grass.t2 as the name in the editor.
What this will do is look at the .t1 and place it in the t1 layer. If it can not find the layer it will be placed into the currently selected layer.