Fixed several cases in which grid snapping was always applied
The Isometric (Staggered) and Hexagonal (Staggered) map renderers do not
have any way of converting between screen and tile coordinates without
rounding to the nearest tile. This caused object positions, polygon node
positions and object sizes to incorrectly snap to tile positions even
when no snapping is enabled.

The actual snapping is still a little broken for staggered maps, but at
least now it no longer happens when you don't want it to.