Fixed Automapping of Tile Objects
Map Objects are removed from automapping when "DeleteTiles" is used.
Code was already there for this but was comparing tile space with map
space.
Tile Objects can be added to Object Layers through automapping.

Tile Objects having a bottom-left origin and zero width and height was
problematic for automapping.
I've added a convenience function MapObject::boundsUseTile() that will
use the tile, if present, to build the boundary for a given Tile Object.
(Perhaps that function could be better named?)
Note: it's possible that the added functionaliy of
MapObject::boundsUseTile() could just be rolled into
MapObject::bounds(), but I didn't want to (potentially) break legacy
code that may have relied on Tile Objects having an empty bounds. For
instance, do we want Tile Objects to be written out to TMX with non-zero
width and height now?