Handle two possible ways in which Export as Image could fail
The most common way in which it failed on 32-bit systems was by running
out of memory and crashing. Now Tiled should show a message telling the
user to either reduce the zoom level or use a 64-bit version of Tiled.
Unfortunately I could not test the fix yet.

The other way it could fail, which is more likely on 64-bit systems, is
that the resulting image would be too big for QImage to handle. In this
case it would generally silently fail, but now it shows an informative
message when this happens.

Closes #907