Changed the way tile probability is applied
Previously, choosing the wrong probabilities could cause the terrain
brush to either fail to find any matches even though there were some or
to never use certain possible matches.

Now, each tile has a default probability of 1 and the probability is
always relative to the other tiles. So a tile with probability 10 is
chosen 10 times as often as a tile with default probability 1, while a
tile with probability of 0.1 is chosen 10 times less.

A probability of 0 can still be used to make sure a certain match is
never chosen.