Changed the look of the resize handles
Now they are arrows, which removed the need for using the different
mouse cursors and adapts correctly to rotated items. The positioning and
look of rotation handles has been tweaked to match.

Also, a change of selection is now avoided when one of the handles was
clicked and the rotation origin indicator is hidden while resizing.