Fixed the grid snapping while resizing
The mAlignPosition was not set when starting the resize, but the
snapping code was relying on it. I decided it did not make that much
sense to even use an alignment position while resizing, and I rather
just snap the handle position to the grid. This seems more
understandable and possibly also more usable than the previous behavior
of snapping the size of the object to the grid.