Allow dragging the origin indicator around before rotate/resize
The position of the origin indicator is now used while rotating as
well as when resizing with Shift modifier.

Currently the position of the origin indicator resets after the
operation. This behavior could still be improved a little. For example
Inkscape seems to remember the relative origin position on a per-object
basis.