Fixed a crash in the terrain brush
If the terrain brush would be used without a terrain selected (can
happen when trying to capture terrain with right mouse button), and it
encounters an empty spot with neighbors for which it chose a certain
terrain tile, it could crash in an attempt to find a new fitting tile
for this spot due to using a null tileset pointer.

Thanks to Thomas Feichtmeir for providing his map for reproducing this
issue!

Probably fixes #222
Probably fixes #372