Fixed updating of map view when changing objects from properties view
The updating of the map view was broken when changing the position, size
or rotation of the object using the properties view.

There's still another problem causing the properties view not to update
correctly after objects are dragged or resized in the map view. Both
problems were caused by the fix in commit 3d2513f5a5dd710094d702b9d3acb.