Fixed updating of Properties view while objects are moved/resized
The events that the updating was relying on where broken in change
3d2513f5a5dd710094d702b9d3acb. Since that change was alright in
principle, the ObjectSelectionTool was changed to emit the changed
signals manually after changing the objects.

Now the updating is even better than before, with the numbers changing
in real time as opposed to only when the mouse is released, as was the
case with Tiled 0.11.