Fixed live automapping updates when moving the mouse fast
The option "Use Automapping, when drawing into layers" was not working
correctly when the stamp brush was used and the mouse was moved fast
while painting.

This was due to wrong edit regions being emitted when fast mouse
movement caused the stamp brush to place the stamp multiple times on a
line.

The automapping could be made a lot faster in this case by only emitting
the edited region once for the whole line, but I'll leave that for
another time.

Fixes #958