Added tilecount property to TMX, JSON and Lua map formats
This can be useful when implementing a map reader, since you will know
in advance how much space to allocate.

If the tileset image changed size since the map file was written, it may
of course contain a different amount of tiles. Care should be taken to
handle that case correctly.

Closes #806