CSV plugin: Some tweaks and core style fixes
The main behavioral change is to keep the old behavior of not changing
the exported file name in the case of maps with a single tile layer,
rather than putting the name of this layer in the file like is still
done when there are multiple tile layers.

Other than that it's mostly some code format changes and tweaks to the
translatable texts.