Fixed removeProperty getting called multiple times on Delete
We should not call removeProperty when handling the ShortcutOverride
event. Accepting this event is just telling Qt that indeed we want to
override the shortcut, and the actual handling of the local shortcut
should happen on the KeyPress event.

This did not actually have any effect, because after the first call
there is no longer a selected property. However, that will change with
pull request #991.