Small stamp preview improvement and worked around repaint issues
While holding the mouse and painting, the stamp preview would not
represent any state that actually got drawn onto the map. This was due
to an optimization to avoid unnecessary work when the mouse is moving
fast.

Now the code may be a bit slower than before, but it's still faster than
previous Tiled versions (very much so when using live automapping) and
it makes sure the preview always shows the right thing.

As a side effect, the repainting issues triggered by the previous
implementation do not seem to happen anymore.