Raised minimum Qt version to 5.1
Raising the version mainly in order to use functionality in recent Qt
versions more comfortably. Also, maintaining compatibility would be
mostly a waste of time since the provided binaries for all systems are
already based on Qt 5 and no serious issues are known.

This also changes some remaining plugins to use QSaveFile instead of
QFile, which avoids writing a truncated file in case of error or running
out of space.