Preserve default file type when doing Save As
There was a bug where the writer plugin file name was used as filter,
which doesn't work. Now, a best effort is made to turn the plugin file
name into a fitting name filter.

The solution will not always work when a single plugin exposes multiple
name filters. This is an issue that will be addressed soon by a larger
patch.

Closes #1049