Improved the plugin system
The previous plugin system worked by direct inheritance from interfaces
like MapWriterInterface and MapReaderInterface. This proved problematic
because it was not possible to implement conflicting interfaces or to
implement a single interface multiple times. The latter affected the
functionality of the Python plugin.

It was solved by having the plugins add objects to the PluginManager when
they are initialized. This way, they can add any number of objects that
implement the available interfaces. In this case, the plugin itself only
extends the new Plugin class.

In addition, MapWriterInterface and MapReaderInterface were merged into a
single MapFormat class. This simplifies implementation of formats that are
both readable and writable. It also makes it easier to save a map in the
same format as it was loaded from.

The changes affect especially the Python plugin in a major way. It is now
possible to implement Python scripts that only support exporting the map.
However, reloading of scripts is (temporarily) not supported.

Closes #897