Qbs: Disable the Python plugin in unsupported configurations
On Windows, the Python plugin does not build in debug mode, nor does it
build with Visual Studio. And of course it won't work if Python is not
installed at all. Now the plugin is automatically disabled in these cases.

In addition the plugin unfortunately does not compile when Python 2.7.10
is installed rather than 2.7.9. I'm not sure about a good way to check
for this, so for now a comment about this will have to suffice.