Fixed issue with editing properties of tiles captured from the map
Only the "current object" was being changed, which left the "selected
tiles" list point to any previous tiles selected in the tileset view,
which were the tiles to which edited properties would actually get
applied.

This mechanism does now provide a way to edit properties of tiles from
multiple tilesets at the same time, which does not handle the
restriction of not being able to edit external tilesets properly. I
think it's rare enough to not need checking right now, but it's a
situation that I aim to resolve before the next feature release.