Load tile images after loading the map
It is still done as part of MapReader::readMap, but making progress
towards handling broken links better.

This change breaks auto-adjustement of tile indexes in the event of a
tileset image becoming wider. This should be reimplement as an optional
thing to do after loading, after asking the user.

A downside of this change is that the map reader can now no longer
report invalid tile references, because it does not know a reference is
invalid until it is too late. This will need to be resolved.