Don't keeps track of tile sizes and offsets in TileLayer
This was an optimization done to keep down rendering overhead, however
when tile images and offsets can change dynamically keeping this
information up-to-date becomes time-consuming since it means we'd often
have to scan all tile layers for which tiles they are using.

As an alternative, a TileLayer now keeps track of which tilesets it is
using. This information only changes when the tile layer itself is
changed, making it much easier to keep up-to-date.