Allow changing tileset image parameters
Implemented via a dialog, reusing the New Tileset dialog. The dialog can
be opened via a small "Edit..." button in the tileset properties view.
Adding this button was sadly the most tricky part of this change...

Changing the tileset parameters can lead to less tiles in the tileset.
Since the tile instances are directly referenced, white placeholders are
kept around when this happens. These should be made more distinguishable
are should probably be hidden in the tileset view.

Closes #331 (Unable to change tileset image source in UI)
Closes #773 (Allow editing tileset parameters)
Closes #507 (Changing transparent color of a tileset image)