Allow changing the image source of individual tiles
This only applies to tiles in an image collection tileset. Broken
references to individual tile images are now no longer raised as an XML
parsing error.

Also, an "image-missing" icon was introduced which shows in the tileset
view for tiles which images could not be loaded, and in the properties
view for when the referenced file isn't found.

Issue #913