Added ability to change the reference to an external tileset
By changing the "Filename" property in the Properties view. This
effectively replaces the tileset.

If the map uses tiles that are not present in the new tileset,
placeholders will be created to avoid the loss of data.

This change is necessary in preparation of allowing the loading of map
files for which some external tileset references could not be found.

Issue #913