Offer to adjust tiles in open maps when tileset width changes
When the tileset image is automatically reloaded, and it turns out that
the number of tile columns in the image changed, it makes sense to ask
the user whether he wants tiles references in open maps to be adjusted.

Unfortunately, at the time the pop-up shows up, the map view will still
be showing the old tiles. This makes it a little hard to decide whether
adjusting is needed. User will be tempted to click "No" but end up with
a messed up map...

Closes #742