Changed layer label in status bar to a combo box
The combo box shows layer type in addition to layer name, and allows
switching between layers.

Also, it fixes a crash that could happen when having layers with very
long names, which was due to the layer name label in the status bar
causing the window to grow very wide.

Fixes #1136