Fixed possible crash when removing layers
When the layer got removed, the LayerModel emitted appropriate signals,
which caused the current index in the new layer combo box to change.
Then Tiled tries to set the current layer based on that, but since the
MapScene had not been notified yet about the layer removal, its attempt
to update the current layer highlight failed with a crash.

In fact we don't want the layer combo box to change the current layer at
all, unless the user actually activated an item there. So use its
"activated" signal instead of the "currentIndexChanged" signal.