Fixed crash in tmxrasterizer and tmxviewer
For maps using external tilesets, the plugin manager is invoked in order
to look for any plugins supporting the file format. But currently the
tmxrasterizer and tmxviewer do not load plugins and do not instantiate a
PluginManager, leading to a null pointer reference.

This change avoids the crash in this case, but it would be interesting
to support plugins also in these applications eventually.