Allow plugins to be enabled/disabled
Project-specific plugins are now disabled by default. All plugins can be
explicitly enabled or disabled in the preferences. This should resolve a
major source of confusion among new users. It is also useful to avoid
extension conflicts, for example there were two plugins that exported to
".lua" files.

Which plugins are enabled currently also affects the --export-map and
--export-formats command line options. Since this can actually be useful
I'll keep it that way for now.

This change also adds feedback in case plugins fail to load. In this
case the error message can be found in the tool tip of the entry in the
preferences dialog.

Closes #1088