Added Object Types editor
This new editor allows not only defining your object types and the color
of objects of that type, but also specifying the default properties for
each type and their default value. This makes setting values for these
properties more efficient and makes it clear what properties an object
of a certain type can take.

Previously specified property types and their colors are preserved and
saved to the new location when changing them.

This change also contains some related fixes:

* Fixed updating of properties view when the object type is changed.
* Fixed updating of properties view when changing the predefined
  properties for the currently selected object.
* Disallowed removing or renaming of predefined properties when in the
  object properties view.

Closes #367