Don't reuse a previous Tiled instance when no arguments are passed
I think it would be entirely reasonable to always reuse an existing
instance, when Tiled would support opening multiple windows. But since
Tiled doesn't, launching the application again is the only way of putting
two maps next to each other. Relying on the --new-instance parameter for
this is too obscure, especially for Windows users.

We can still support the file manager use-case, as well as the use-case of
opening extra maps from the command line, by reusing a previous instance
in case any files are passed on the command line.