Windows installer: Removed the "[ProductName]" start menu folder
Apparently this value substitution does not work for that field for some
reason. The Tiled executable can just sit at the top level anyway, since
it's so lonely otherwise.

Also switched the registry key from HKCU to HKMU, which is supposed to
adapt to single-user or multi-user installations. However I don't think
the installer currently provides the user any choice in that.