Optimized selection of many objects
When selecting many objects at the same time, Tiled was getting very slow
to respond to selection changes. This was caused for a large part by the
way the Objects view was synchronizing its selected items.

Now, instead of calling QItemSelectionModel::select for each selected
object, it is called only once with the complete selection. In my
measurements with 500 selected objects, this took the execution time of
this function down from about 1300ms to about 300ms.

This change also fixes an even more serious problem when switching maps
with many objects selected. This lead to long unresponsive periods due
forgetting to set mSynching to true.