Sparkle: Display UI on manual update check
When the user manually checks for updates, there should be a progress
bar and a message even if there is no update found.

To get the update check to actually work in the Preferences dialog, the
dialog could not be application modal. I don't think the Preferences
dialog actually needed to be modal, anyway.

Unfortunately, Qt on OS X has a bug that allows child windows to
disappear behind their parent window. So now the Preferences dialog can
go behind the main Tiled window...