Fixed assertion on exit when keeping preferences open
The PreferencesDialog destructor was re-instantiating the Preferences
singleton, which in turn re-instantiated the TilesetManager and
PluginManager singletons. Then, when the plugins got destroyed, they
were trying to remove their objects from the newly instantiated
PluginManager, but then it failed an assertion because it didn't know
about these objects.

Lesson learned: singletons are tricky beasts, and accessing them from
destructors is especially dangerous.