Fixed potential crash and too much switching to Terrain Brush
When switching maps and the new map did not have a tile layer selected,
the Terrain Brush ran into an assertion failure because it was not
immediately getting deactivated.

Also, the automatic switching to the Terrain Brush was too aggressive,
leading to this tool being selected by default when any terrain was
present.