Qbs: Fixed build issue related to Info.plist file
For now, just moved the Info.plist file to the 'osx' section. I have not
checked yet whether it actually does what is expected there.

This is happening with more recent versions of Qbs:

error: Cannot install '.../tiled/src/tiled/Info.plist', because it doesn't start with the value of qbs.installSourceBase '...'.