Qbs: Added generation of 7-zip archive for Windows
It is not built by default. Instead it can be triggered by "-p archive" or
"--all-products" (this one will also build the installer).

Also configured AppVeyor to upload the archive and synchronized the
location of the Windows snapshots with the OS X snapshots.

Related to issue #366, but right now this is not a special version
regarding where settings are stored and it does still include the update
check, which will download the MSI installer. Especially that last thing
should be changed before this can be distributed (for example through the
itch.io app).