Fixed updating object drag cursor when exiting resize handles
We can't use item-specific mouse cursors in combination with manually
updating the viewport mouse cursor based on graphics scene mouse events,
because of the way QGraphicsView saves and restores the viewport cursor.

In addition, we could no longer rely on the default drag functionality
for the origin indicator because it would grab the mouse in case it was
below another handle. So now a "MovingOrigin" action was added that
implements that part.

Closes #1277