Automapping: Fixed a painting related bug affecting the top edge
When operations done by the AutoMapper affected the draw margins of a
tile layer (like, by putting tiles on a layer for the first time), the
TileLayerItem was not informed as necessary for it to take on the
correct bounding rectangle. This could lead to the top row of tiles to
not be painted at all.

This change also includes some small code cleanups, and an optimization
that avoids copying all affected layers twice.

Closes #1308