Tiled Fusion: Darker text color for unselected tabs
When the base color is dark, the text is light. Since the unselected tabs are
rendered darker, this means the text has a stronger contrast there. This is
avoided by darkening the text on these tabs.

For the light version this is not needed, because dark text already has a bit
less contrast on the darker tabs.