Correctly preserve tile meta-data when tileset width is changed
When the tileset image width changes, affecting the number of tile
columns in the tileset, the user is asked whether the tile references in
the map should be adjusted accordingly. However, the meta-data stored
for each tile, like collision information, properties and animations,
was not being adjusted, leading to corrupted data.

This change extends the behavior when tile indexes get adjusted, to make
sure also tile meta-data and tile references from tile animations get
adjusted as appropriate.

Closes #1315