OS X: Use standard shortcut for toggling full screen
F11 can't be used since it's a window manager shortcut. The standard
shortcut is Ctrl+Cmd+F.

There is actually also an automatically added "Enter Full Screen" action
visible in the "View" menu. It would be nice to set the shortcut on that
one instead and not add our own, but currently I don't know where this
action comes from.