Avoid using direct Tile pointers in tile layers and objects
Using direct Tile pointers has several downsides. For example, it is
hard to make sure they stay valid when tiles are removed from a tileset.
And when doing so, it could lead to data loss, for example when tiles
temporarily disappear when a tileset definition is changed or its image
is resized. Currently, tile instances are being kept alive because there
might still be pointers to them out there.

Instead, the Cell now uses a combination of Tileset pointer and tile ID.
A lot of code has now been adjusted so that it does not assume that a
non-empty Cell means that a valid Tile instance can be found.

In addition, when an invalid tile is encountered while rendering the
map, a noticeable red marker is rendered instead.

There are still some Tile pointers used in the code. Depending on their
lifetime, I may need to introduce a 'TileRef' struct that combines a
Tileset pointer with a tile ID, to replace those Tile pointers.