Avoid saving unfinished objects
When pressing Save (Ctrl+S) while adding a new object to the map, the
new object would already be saved as part of the map, even though its
creation can still be canceled (right-click or Escape) or it may still
be an incomplete polygon or polyline.

Prevent this by adding the new object to a dummy object group instead.

Closes #1348