tmxrasterizer: Use smooth pixmap transform by default
This is to increase the default quality of the rendered map, especially
when rendering thumbnails. It can be turned off separately from anti-
aliasing using the new --no-smoothing parameter.

Anti-aliasing remains off by default since it is usually not what you
want. It applies anti-aliasing to edges of primitives, including tiles.
This creates visible seems at tile boundaries when the tiles are not
pixel-aligned.