Initial start page when no files are open
This change adds a special page that is shown when no files are open.
It currently just helps the user to get started by providing buttons to
either create a new map or a new tileset.

It also adds the notion of globally registered actions, which is one way
of allowing this new page to trigger the "New Map" and "New Tileset"
actions. The Id and the ActionManager are loosely based on code in Qt
Creator, though I hope to keep things a bit more simple. It would be
nice if eventually these registered actions could feature for example
custom shortcuts.