Immediately ask to save new maps and tilesets
It will no longer be allowed to create some new unsaved file, because
this doesn't work nicely when other files are saved that reference the
file. This particular problem currently only applies when saving maps
that reference some unsaved tileset, but the behavior will be the same
for consistency.

Also fixed a small bug causing the extension mismatch warning to pop up
when canceling the Save As dialog.