Tweaked the tool bar style to fit better with the tabs
No more light line at the top and the gradient no longer starts lighter
than the color of the tab.

I'd probably have left the line there for any other tool bar lines than
the first, but I can't because of what appears to be a bug in Qt regarding
setting the QStyleOptionToolBar::positionOfLine member.