Fixed crash when saving a new tileset for the second time
On the second save, the reload mechanism was kicking in and it was
crashing because the reader format was never set.

First of all the reload mechanism shouldn't kick in. This was due to
forgetting to set the last saved time. And now when no reader format is
set, it relies on the fact that there must be a writer format set that
can also read.