Tileset margin and spacing for gmx plugin
Hey @bjorn and @JonesBlunt,

Thanks for making this plugin.

Not sure if it was intended or not, to not include margin and spacing, so I added few code lines to support that.

Also wanted to ask what was the reason to leave tile->tileset()->tileWidth() for yo? Changed to tile->tileset()->tileHeight().