Introduced a ReversingProxyModel
This was way more work than I had anticipated, but this proxy model
simplifies the LayerModel and a few other places that needed to take
into account that the layers were displayed in reversed order.

It will also be used to apply the same reversed order to the Objects
view.