Support Paste in Place action for tile layers
Now, when a tile selection was copied and a tile layer is selected, the
Paste in Place action will directly paste the copied tile selection into
the current layer, at the same position where it was copied from.

This provides a somewhat convenient way of moving a region of tiles from
one layer to another.

Closes #1415