MSI Installer: Fixed the architecture of shipped MSVC DLLs
The 64-bit Tiled version was shipping 32-bit MSVC DLLs, and vica versa,
due to me being confused by the directory names.

This issue was only noticable on relatively clean systems, because as soon
as the MSVC DLL was installed as a dependency of some other application
Tiled would anyway start up fine.

Thanks to reedb for noticing the mixup and posting about it on the Tiled
forum! http://discourse.mapeditor.org/t/help-tiled-broke/1851/3