Made safe writing of files optional
Since the safe (atomic) writing of files is unfortunately causing
problems on network drives and in Dropbox folders, there is now an
option to disable it entirely.

Users should be aware that disabling it will result in partially written
files in the event of Tiled crashing while saving or when there is
insufficient space available. When overwriting, the previous contents of
the file will become unavailable, so backups should be made.

Closes #1402
Closes #1404