Fixed crash when resizing map causes objects to get removed
This bug sneaked in due to a QUndoCommand getting immediately executed
even when pushed as part of a macro, whereas the code was relying on it
getting executed later (in using a modern for-loop that assumed the
iterated container is not going to get modified).

I expected execution to be later, since the beginMacro/endMacro are
documented with a piece of "equivalent" code that would behave like
that. Turns out it's not exactly the same, and I was a fool for assuming
it was.