Allow setting up the orientation and grid size of a tileset
This orientation (currently just isometric or orthogonal) allows the
collision editor and terrain overlays to take into account isometric
tiles and transform the rendering and mouse input accordingly.

Useful for when the collision shape should lie "on the ground" and to
better see which corner is meant regarding the terrain information.

Closes #757 - Collision Objects are not projected while in Isometric Mode
Closes #419 - isometric terrain overlays