Fixed encoding problems with command-line output
With some languages on some systems, much of the command-line output
would be unreadable. This was seen on Windows with Chinese language. It
was likely caused by passing the result of qPrintable, which produces a
string with local 8-bit encoding, to qWarning/qDebug/etc, which expect a
UTF-8 encoded string. The problem was not generally visible on Linux,
where the local 8-bit encoding is usually UTF-8.

The code was now changed to use qUtf8Printable where appropriate, as
well as QDebug::noquote, because most of the old qPrintable usages were
to avoid the quotes.

qUtf8Printable and QDebug::noquote were added in Qt 5.4, so this is now
the minimum supported version of Qt (previously 5.2).

Closes #1381