Added GroupLayer and GroupLayerItem
A GroupLayer is a new kind of layer that can contain other layers. It is
meant for organizing layers in folders in case you are working with maps
that have a large amount of layers.

The visibility, opacity and later also locked status of the group layer
will affect all child layers. Also the Layers view will allow collapsing
group layers.

This change is only the start. The addition of group layers is expected
to affect a large amount of code.