Improved DPI-awareness in many places
Tiled now tries to adapt to the DPI of the primary screen as much as
possible. This way, we can hopefully stop relying on full application
scaling as a compatibility solution. Using a global scale has several
downsides, including painting bugs and overly fat lines.

The following things now adapt to the DPI of the primary screen (needing
restart if it changes):

* Hard-coded icon sizes
* Initial window sizes
* Rounded corners in Tiled Fusion style
* Size hints of scroll areas
* Mouse wheel scroll speed on map view (though it still seems too slow)
* Resize and rotate handles

The main thing remaining is to provide higher resolution versions of
most icons.