Added diagonal mirroring mode to terrain brush
When holding the Alt modifier while using the terrain brush, the
modifications will be mirrored diagonally (top-left / bottom-right).

In general, there could also be horizontal or vertical mirroring as well
as a combination of those, but that will require adding a tool-specific
configuration panel somewhere (will have to be done eventually).

Thanks to 'Peter' for pushing for this feature.