Terrain Brush: Changed the way the mirroring works
Previously a crash could occur when using the Alt "mirror" modifier of
the Terrain Brush on a non-square map. The new mirroring approach is
actually not technically behaving like a mirror, but it works also for
rectangular maps and I think in general it yields nicer looking maps.
It's actually doing a 180 degree rotation rather than a mirror.