Avoid potential issues with initializing object IDs
When child ObjectGroup layers would be added to a GroupLayer instance
before it is added to the map, the objects in the ObjectGroup would not
get their IDs initialized.

Delegate the setting of map ownership to the Map class, so that the IDs
will initialize also in such a scenario.