Added a Text object
This Text object has already many features, including word wrapping,
coloring and font options. The main thing lacking still is a UI to
change its vertical and horizontal alignment properties.

Another thing that may be useful, is to have an option to adjust the
size to the contents, rather than the other way around. Currently, the
text is wrapped and clipped to the size of the object, which is not the
most convenient thing depending on the use-case. It can also cause
problems when different native fonts are used or font rendering differs
slightly between systems.

Closes #1429