Added Tile.type and inherit tile object properties from the tile
This type is the same as used for objects, and tile objects without
explicit type will inherit it from their tile. When the type is
inherited, it is displayed in gray (in Properties and Objects view).

Custom properties set on the tile are also inherited by tile objects
using that tile.

Closes #436
Closes #1248