Fixed Offset Layers tool to wait with dragging
The Offset Layers tool caused an undo entry to be made by just clicking,
because it was immediately going into dragging mode when the mouse was
pressed. Now it waits until the mouse has been moved by at least half
the QApplication::startDragDistance(), before it will start the drag
operation.

Also fixed an issue where using a shortcut to switch to another tool
while dragging was preventing the creation of an undo entry for the
drag.