Always update current stamp when a tile is clicked
Regardless of whether it changed the selection or not. This makes sure
that the stamp is updated when clicking an already selected tile in the
tileset view, which can be the case if the stamp was previously a stamp
with the same tile repeated, or when the stamp contained tiles from
other tilesets.