Improved the behavior of the tile selection tool
It is now easier to select tiles since it works the same way as when
capturing a stamp. Any touched tile is included in the selection. The
only exception remains when you simply click, which clears the selection
because selecting now only starts when the mouse has been moved at least
a little bit.

Based on feedback from @eevee.