Fixed error reported by valgrind
The construction of mZoomable causes an event to be delived to the
TilesetView (probably QEvent::ChildAdded), causing the event override to
read the pointer before it was initialized.

It could not actually result in any issue since the event type was also
checked, but to keep valgrind silent it's better to reverse this
condition.